Updated for python 3.12 with Ubuntu 24.04

+++ venv_create.sh
@@ -1,6 +1,6 @@
 #! /bin/bash -e
 
-python3 -m pip install virtualenv
+sudo apt install python3-virtualenv
 if [ -d "venv" ];
 then
   sudo rm -rf venv
@@ -10,6 +10,9 @@ sudo chown -R $USER:$USER ./
 python3 -m virtualenv venv
 source venv/bin/activate
 pip install -U pip setuptools wheel
-pip install -r requirements.txt
+if [ -f "requirements.txt" ];
+then
+  pip install -r requirements.txt
+fi
 
 mkdir -p log
